firebase_functions/functions/requirements.txt
firebase_functions~=0.1.0
google-cloud-aiplatform[adk,agent_engines]
google-cloud-documentai
requests
